Added executable file for linux users
